+++ src/GRA.Domain.Repository/IGroupInfoRepository.cs
@@ -1,12 +1,15 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using GRA.Domain.Model;
+using GRA.Domain.Model.Filters;
 
 namespace GRA.Domain.Repository
 {
     public interface IGroupInfoRepository : IRepository<Model.GroupInfo>
     {
-        Task<IEnumerable<Model.GroupInfo>> GetAllAsync(int siteId);
+        Task<IEnumerable<GroupInfo>> GetAllAsync(int siteId);
         Task<int> GetCountByTypeAsync(int groupTypeId);
-        Task<Model.GroupInfo> GetByUserIdAsync(int householdHeadUserId);
+        Task<GroupInfo> GetByUserIdAsync(int householdHeadUserId);
+        Task<DataWithCount<ICollection<GroupInfo>>> PageAsync(GroupFilter filter);
     }
 }