+++ src/GRA.Domain.Repository/IGroupTypeRepository.cs
@@ -6,6 +6,7 @@ namespace GRA.Domain.Repository
 {
     public interface IGroupTypeRepository : IRepository<GroupType>
     {
+        Task<GroupType> GetDefaultAsync(int siteid);
         Task<IEnumerable<GroupType>> GetAllForListAsync(int siteId);
         Task<(IEnumerable<GroupType>, int)> GetAllPagedAsync(int siteId, int skip, int take);
     }